Call setup_fairseq2 automatically
