+++ src/fairseq2/cli/_main.py
@@ -13,14 +13,14 @@ from signal import SIG_DFL, SIGINT, raise_signal, signal
 import torch
 from torch.cuda import OutOfMemoryError
 
+from fairseq2 import setup_fairseq2
 from fairseq2.cli._logging import setup_logging
 from fairseq2.cli._setup import setup_cli
 from fairseq2.cli.utils.rich import create_rich_progress_reporter
-from fairseq2.context import get_runtime_context
 from fairseq2.error import ContractError, InternalError
 from fairseq2.extensions import ExtensionError
 from fairseq2.logging import LoggingSetupError, log
-from fairseq2.setup import SetupError, setup_fairseq2
+from fairseq2.setup import SetupError
 from fairseq2.utils.env import InvalidEnvironmentVariableError, get_rank
 
 
@@ -68,9 +68,7 @@ def _run() -> int:
 
         progress_reporter = create_rich_progress_reporter(rank)
 
-        setup_fairseq2(progress_reporter)
-
-        context = get_runtime_context()
+        context = setup_fairseq2(progress_reporter)
 
         cli = setup_cli(context)
     except SetupError:
