+++ src/fairseq2/models/_hub.py
@@ -11,6 +11,7 @@ from typing import Generic, TypeVar, cast, final
 import torch
 from torch.nn import Module
 
+from fairseq2 import get_runtime_context
 from fairseq2.assets import (
     AssetCard,
     AssetCardError,
@@ -18,7 +19,6 @@ from fairseq2.assets import (
     AssetCardNotFoundError,
     AssetStore,
 )
-from fairseq2.context import get_runtime_context
 from fairseq2.gang import Gangs, fake_gangs
 from fairseq2.models._error import (
     InvalidModelConfigTypeError,
