+++ src/fairseq2/setup/__init__.py
@@ -12,7 +12,6 @@ from fairseq2.setup._asset import (
 from fairseq2.setup._datasets import DatasetRegistrar as DatasetRegistrar
 from fairseq2.setup._error import SetupError as SetupError
 from fairseq2.setup._models import ModelRegistrar as ModelRegistrar
-from fairseq2.setup._root import setup_fairseq2 as setup_fairseq2
 from fairseq2.setup._root import setup_library as setup_library
 from fairseq2.setup._text_tokenizers import (
     TextTokenizerRegistrar as TextTokenizerRegistrar,
